+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-01-10] [Export layout](https://github.com/RubricLab/ui/commit/27dbfeeb8d9b3a990699f17b7e53448b4e15a71f)
 - [2025-01-10] [Improve charts](https://github.com/RubricLab/ui/commit/0aea72e666e2ecef154e9b4909952d07d7eb4511)
 - [2025-01-10] [Add line and bar chart](https://github.com/RubricLab/ui/commit/8de4cd8b0086f48d4116b96cc859f947ad6b629b)
 - [2025-01-10] [Make tag look better](https://github.com/RubricLab/ui/commit/822f862f93645c939c903ddf1a62c7b3e6ca179a)